constructive_final/requirement.txt
joblib
pandas
scikit-learn
NRCLex
pyenchant
streamlit
scikeras
tensorflow
NormalizeContraction
bs4
matplotlib
wordsegment
gensim